rename highlighter to replacer
