+++ src/Replacer/Highlighter/Highlighter.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace VStelmakh\UrlHighlight\Highlighter\Linker;
+namespace VStelmakh\UrlHighlight\Replacer\Highlighter;
 
 use VStelmakh\UrlHighlight\Matcher\UrlMatch;
 
-interface Linker
+interface Highlighter
 {
     public function render(UrlMatch $match): string;
 }
