+++ src/Replacer/Tokenizer/Token/CommentToken.php
@@ -2,18 +2,15 @@
 
 declare(strict_types=1);
 
-namespace VStelmakh\UrlHighlight\Highlighter\Tokenizer\Token;
+namespace VStelmakh\UrlHighlight\Replacer\Tokenizer\Token;
 
 /**
  * @internal
  */
-final readonly class TagToken implements Token
+final readonly class CommentToken implements Token
 {
     public function __construct(
         public string $contents,
-        public string $name,
-        public bool $isClosing,
-        public bool $isSelfClosing,
     ) {}
 
     #[\Override]
