+++ src/Replacer/Tokenizer/Token/TagToken.php
@@ -2,15 +2,18 @@
 
 declare(strict_types=1);
 
-namespace VStelmakh\UrlHighlight\Highlighter\Tokenizer\Token;
+namespace VStelmakh\UrlHighlight\Replacer\Tokenizer\Token;
 
 /**
  * @internal
  */
-final readonly class CommentToken implements Token
+final readonly class TagToken implements Token
 {
     public function __construct(
         public string $contents,
+        public string $name,
+        public bool $isClosing,
+        public bool $isSelfClosing,
     ) {}
 
     #[\Override]
